+++ packages/@headlessui-react/src/hooks/use-owner.ts
@@ -1,6 +1,10 @@
 import { useMemo } from 'react'
-import { getOwnerDocument } from '../utils/owner'
+import { getOwnerDocument, getRootNode } from '../utils/owner'
 
 export function useOwnerDocument(...args: Parameters<typeof getOwnerDocument>) {
   return useMemo(() => getOwnerDocument(...args), [...args])
+}
+
+export function useRootDocument(...args: Parameters<typeof getRootNode>) {
+  return useMemo(() => getRootNode(...args), [...args])
 }