+++ packages/@headlessui-react/src/utils/focus-management.ts
@@ -2,7 +2,7 @@ import type { MutableRefObject } from 'react'
 import { disposables } from './disposables'
 import * as DOM from './dom'
 import { match } from './match'
-import { getOwnerDocument } from './owner'
+import { getActiveElement, getOwnerDocument, getRootNode } from './owner'
 
 // Credit:
 //  - https://stackoverflow.com/a/30753870
@@ -86,7 +86,11 @@ enum Direction {
   Next = 1,
 }
 
-export function getFocusableElements(container: HTMLElement | null = document.body) {
+interface QuerySelectorAll {
+  querySelectorAll<E extends Element = Element>(selectors: string): NodeListOf<E>
+}
+
+export function getFocusableElements(container: QuerySelectorAll | null = document.body) {
   if (container == null) return []
   return Array.from(container.querySelectorAll<HTMLElement>(focusableSelector)).sort(
     // We want to move `tabIndex={0}` to the end of the list, this is what the browser does as well.
@@ -136,12 +140,13 @@ export function isFocusableElement(
 }
 
 export function restoreFocusIfNecessary(element: HTMLElement | null) {
-  let ownerDocument = getOwnerDocument(element)
   disposables().nextFrame(() => {
+    let activeElement = getActiveElement(element)
+
     if (
-      ownerDocument &&
-      DOM.isHTMLorSVGElement(ownerDocument.activeElement) &&
-      !isFocusableElement(ownerDocument.activeElement, FocusableMode.Strict)
+      activeElement &&
+      DOM.isHTMLorSVGElement(activeElement) &&
+      !isFocusableElement(activeElement, FocusableMode.Strict)
     ) {
       focusElement(element)
     }
@@ -219,8 +224,12 @@ export function sortByDomNode<T>(
   })
 }
 
-export function focusFrom(current: HTMLElement | null, focus: Focus) {
-  return focusIn(getFocusableElements(), focus, { relativeTo: current })
+export function focusFrom(
+  current: HTMLElement | null,
+  focus: Focus,
+  container = current === null ? document.body : getRootNode(current)
+) {
+  return focusIn(getFocusableElements(container), focus, { relativeTo: current })
 }
 
 export function focusIn(
@@ -236,11 +245,11 @@ export function focusIn(
     skipElements: (HTMLElement | MutableRefObject<HTMLElement | null>)[]
   }> = {}
 ) {
-  let ownerDocument = Array.isArray(container)
+  let root = Array.isArray(container)
     ? container.length > 0
-      ? container[0].ownerDocument
+      ? getRootNode(container[0])
       : document
-    : container.ownerDocument
+    : getRootNode(container)
 
   let elements = Array.isArray(container)
     ? sorted
@@ -262,7 +271,7 @@ export function focusIn(
     )
   }
 
-  relativeTo = relativeTo ?? (ownerDocument.activeElement as HTMLElement)
+  relativeTo = relativeTo ?? (root?.activeElement as HTMLElement)
 
   let direction = (() => {
     if (focus & (Focus.First | Focus.Next)) return Direction.Next
@@ -305,7 +314,7 @@ export function focusIn(
 
     // Try the next one in line
     offset += direction
-  } while (next !== ownerDocument.activeElement)
+  } while (next !== getActiveElement(next))
 
   // By default if you <Tab> to a text input or a textarea, the browser will
   // select all the text once the focus is inside these DOM Nodes. However,